Fix #248: Improve isolation enrollment from onboarding
Avoid initialization of these beans at enrollment server

- WultraMockDocumentVerificationProvider
- WultraMockPresenceCheckProvider
