+++ enrollment-server-onboarding/src/main/java/com/wultra/app/onboardingserver/configuration/PowerAuthWebServiceConfiguration.java
@@ -38,8 +38,8 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 @ConfigurationProperties("ext")
 @ComponentScan(basePackages = {
-        "com.wultra.app.docverify",
-        "com.wultra.app.presencecheck",
+        "com.wultra.app.onboardingserver.docverify",
+        "com.wultra.app.onboardingserver.presencecheck",
         "com.wultra.security.powerauth",
         "io.getlime.security.powerauth",
 })
