+++ enrollment-server-onboarding/src/main/java/com/wultra/app/onboardingserver/docverify/zenid/ZenidConst.java
@@ -1,0 +1,37 @@
+/*
+ * PowerAuth Enrollment Server
+ * Copyright (C) 2022 Wultra s.r.o.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.wultra.app.onboardingserver.docverify.zenid;
+
+/**
+ * Constants for ZenID purposes
+ *
+ * @author Lukas Lukovsky, lukas.lukovsky@wultra.com
+ */
+public class ZenidConst {
+
+    /**
+     * Verification SDK initialization response
+     */
+    public static final String SDK_INIT_RESPONSE = "zenid-sdk-init-response";
+
+    /**
+     * Verification SDK initialization token
+     */
+    public static final String SDK_INIT_TOKEN = "sdk-init-token";
+
+}
