+++ enrollment-server-onboarding/src/main/java/com/wultra/app/onboardingserver/docverify/zenid/model/deserializer/CustomOffsetDateTimeDeserializer.java
@@ -1,0 +1,51 @@
+/*
+ * PowerAuth Enrollment Server
+ * Copyright (C) 2021 Wultra s.r.o.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.wultra.app.onboardingserver.docverify.zenid.model.deserializer;
+
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonDeserializer;
+import com.fasterxml.jackson.datatype.jsr310.deser.InstantDeserializer;
+
+import java.io.IOException;
+import java.time.LocalDate;
+import java.time.OffsetDateTime;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
+
+/**
+ * Custom {@link java.time.OffsetDateTime} deserializer which allows also iso date format
+ *
+ * <p>
+ *     The ZenID returns simple ISO data on some date elements which are expected to be date-time (e.g. BirthDate)
+ * </p>
+ */
+public class CustomOffsetDateTimeDeserializer extends JsonDeserializer<OffsetDateTime> {
+
+    @Override
+    public OffsetDateTime deserialize(JsonParser parser, DeserializationContext context) throws IOException {
+        if (parser.getText() != null && parser.getText().length() == 10) { // yyyy-MM-dd
+            return LocalDate.parse(parser.getText(), DateTimeFormatter.ISO_DATE)
+                    .atStartOfDay(ZoneId.of("UTC"))
+                    .toOffsetDateTime();
+        }
+
+        return InstantDeserializer.OFFSET_DATE_TIME.deserialize(parser, context);
+    }
+
+}
