--- enrollment-server-onboarding/src/main/java/com/wultra/app/presencecheck/iproov/IProovConst.java
@@ -1,32 +1,0 @@
-/*
- * PowerAuth Enrollment Server
- * Copyright (C) 2021 Wultra s.r.o.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package com.wultra.app.presencecheck.iproov;
-
-/**
- * Constants for iProov purposes
- *
- * @author Lukas Lukovsky, lukas.lukovsky@wultra.com
- */
-public class IProovConst {
-
-    /**
-     * Session parameter name of the verification token
-     */
-    public static final String VERIFICATION_TOKEN = "iproov-verification-token";
-
-}
